transfermarkt-api/pyproject.toml

[tool.poetry.dependencies]
python = "^3.12"
redis = "^5.2.1"
openai = "^1.58.1"
deepseek = "^0.0.1.dev0"
python-dotenv = "^1.0.1"
httpx = "0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
authlib = "^1.4.0"
fastapi-auth0 = "^0.5.0"
pyjwt = "^2.10.1"
firebase-admin = "^6.6.0"
fuzzywuzzy = "^0.18.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
selenium = "4.26.1"
slowapi = "^0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
